requirements.txt
discord.py==2.3.2
motor==3.3.2
python-dotenv==1.0.0
matplotlib==3.8.2
pymongo==4.6.1
aiohttp==3.9.1